Move sort plugins to parent
