+++ packages/plugin/src/index.ts
@@ -1,7 +1,7 @@
 export * from './PluginsContext';
 export * from './PluginTypes';
 export * from './PluginUtils';
-export * from './pluginSortUtil';
+export * from './sortPluginsByDependency';
 export * from './TablePlugin';
 export * from './useCustomThemes';
 export * from './useDashboardPlugins';
