--- packages/plugin/src/pluginSortUtil/index.ts
@@ -1,1 +1,0 @@
-export * from './sortPluginsByDependency';
